requirements.txt
docling>=0.1.0
python-dotenv>=1.0.0
litellm>=1.0.0 